+++ app/controllers/follows_controller.rb
@@ -10,12 +10,12 @@ class FollowsController < ApplicationController
       return
     end
 
-    following_them_check = CheckCachedFollows.call(current_user, params[:followable_type], params[:id])
+    following_them_check = Follows::CheckCached.call(current_user, params[:followable_type], params[:id])
 
     return render plain: following_them_check unless params[:followable_type] == "User"
 
-    following_you_check = CheckCachedFollows.call(User.find_by(id: params[:id]), params[:followable_type],
-                                                  current_user.id)
+    following_you_check = Follows::CheckCached.call(User.find_by(id: params[:id]), params[:followable_type],
+                                                    current_user.id)
 
     if following_them_check && following_you_check
       render plain: "mutual"
@@ -34,9 +34,9 @@ class FollowsController < ApplicationController
       if current_user.id == id
         "self"
       else
-        following_them_check = CheckCachedFollows.call(current_user, params[:followable_type], id)
-        following_you_check = CheckCachedFollows.call(User.find_by(id: id), params[:followable_type],
-                                                      current_user.id)
+        following_them_check = Follows::CheckCached.call(current_user, params[:followable_type], id)
+        following_you_check = Follows::CheckCached.call(User.find_by(id: id), params[:followable_type],
+                                                        current_user.id)
         if following_them_check && following_you_check
           "mutual"
         elsif following_you_check
